Split up core into core-logging / core-utils
