+++ packages/core-logging/.gitignore
