+++ packages/core-logging/.prettierrc
