+++ packages/core-logging/jest.config.js
