+++ packages/core-logging/src/errorHandler.ts
