+++ packages/core-logging/src/index.ts
@@ -1,4 +1,3 @@
 export * from './errorHandler'
 export * from './models'
 export * from './logger'
-export * from './utils'
