+++ packages/core-logging/src/logger.ts
@@ -1,6 +1,6 @@
 import { sprintf } from 'sprintf-js'
 import { getErrorHandler } from './errorHandler'
-import { isValidEnum } from './utils'
+import { isValidEnum } from '@optimizely/js-sdk-core-utils'
 
 import { LogLevel, LoggerFacade, LogManager, LogHandler } from './models'
