+++ packages/core-logging/src/models.ts
