+++ packages/core-utils/.gitignore
@@ -1,0 +1,2 @@
+lib/
+doc/
