+++ packages/core-utils/jest.config.js
@@ -1,0 +1,17 @@
+module.exports = {
+  // "roots": [
+  //   "./src"
+  // ],
+  "transform": {
+    "^.+\\.tsx?$": "ts-jest"
+  },
+  "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$",
+  "moduleFileExtensions": [
+    "ts",
+    "tsx",
+    "js",
+    "jsx",
+    "json",
+    "node"
+  ],
+}
