+++ packages/core-utils/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@optimizely/js-sdk-core",
-  "version": "0.6.0",
-  "description": "Optimizely Full Stack Core",
+  "name": "@optimizely/js-sdk-core-utils",
+  "version": "0.1.0",
+  "description": "Optimizely Full Stack Core Utils",
   "author": "jordangarcia <jordan@optimizely.com>",
   "homepage": "https://github.com/optimizely/javascript-sdk",
   "license": "MIT",
@@ -20,12 +20,10 @@
     "prepublishOnly": "jest && npm run tsc"
   },
   "dependencies": {
-    "sprintf-js": "^1.1.1",
     "uuid": "^3.3.2"
   },
   "devDependencies": {
     "@types/jest": "^23.3.12",
-    "@types/sprintf-js": "^1.1.1",
     "@types/uuid": "^3.4.4",
     "jest": "^23.6.0",
     "ts-jest": "^23.10.5"
