+++ packages/optimizely-sdk/LGr2Yv4rochGYCvxRCPZ1d.json
@@ -1,0 +1,1 @@
+{"version": "4", "rollouts": [], "typedAudiences": [], "anonymizeIP": false, "projectId": "13627400256", "variables": [], "featureFlags": [], "experiments": [{"status": "Running", "audienceIds": [], "variations": [{"variables": [], "id": "13625360250", "key": "var1"}, {"variables": [], "id": "13648280126", "key": "var2"}], "id": "13623600328", "key": "abtest1", "layerId": "13619500363", "trafficAllocation": [{"entityId": "13625360250", "endOfRange": 5000}, {"entityId": "13648280126", "endOfRange": 10000}], "forcedVariations": {}}], "audiences": [{"conditions": "[\"or\", {\"match\": \"exact\", \"name\": \"$opt_dummy_attribute\", \"type\": \"custom_attribute\", \"value\": \"$opt_dummy_value\"}]", "id": "$opt_dummy_audience", "name": "Optimizely-Generated Audience for Backwards Compatibility"}], "groups": [], "attributes": [{"id": "13642620365", "key": "gender"}], "botFiltering": false, "accountId": "804231466", "events": [], "revision": "5"}
